src/br/com/alura/rh/service/ReajusteService.java
package br.com.alura.rh.service;

import java.math.BigDecimal;
import java.util.List;

import br.com.alura.rh.model.Funcionario;
import lombok.AllArgsConstructor;

@AllArgsConstructor
public class ReajusteService {
	private List<ValidacaoReajuste> validacoes;

	public void reajustarSalariodoFuncionario(Funcionario funcionario, BigDecimal aumento) {
		this.validacoes.forEach(v -> v.validar(funcionario, aumento));
		BigDecimal salarioReajustado = funcionario.getSalario().add(aumento);
		funcionario.atualizarSalario(salarioReajustado);
	}
}
